predictors/requirements.txt
ipython
pytest
transformers
pytorch-lightning 
matplotlib
numpy
huggingface_hub
torchvision
scikit-learn
pandas
openpyxl